Remove consult-lsp-marginalia
The recent breaking change in consult reminded me that I won't have the
time or the motivation to maintain this glue code that I don't use. The
file still exists in the history though, so it can be used as a starting
point if anyone wants to keep using/tweaking it.