fix(file-symbols): use end-line, stop assuming ht
When symbols span multiple lines, `cend` might be less than `cbeg`, so
we must use the correct line of the symbol instead of assuming the
symbol only lives on one line.

Also, `consult-lsp--flatten-document-symbols` assumed that the return
values from LSP requests was always a hash-table, so in the same vein
as #30 we are changing the code to use lsp-mode-provided helpers.

Closes: #33